Add: rename rooms router
