+++ app.js
@@ -35,8 +35,7 @@ app.use(require("./src/middleware/limitRate"));
 app.use("/auth", require("./src/route/auth"));
 app.use("/json/logininfo", require("./src/route/logininfo"));
 app.use("/users", require("./src/route/users"));
-app.use("/rooms/v2", require("./src/route/rooms.v2"));
-app.use("/rooms", require("./src/route/rooms.v2"));
+app.use("/rooms", require("./src/route/rooms"));
 app.use("/chats", require("./src/route/chats"));
 app.use("/locations", require("./src/route/locations"));
 app.use("/reports", require("./src/route/reports"));
